fixed minor errors
