+++ client/src/components/ProfileView/GitHubLink/index.jsx
@@ -3,7 +3,7 @@ import React from "react";
 const GitHubLink = () => {
   return (
    
-   <a href="" stlye={{textAlign: 'right'}}> GitHubLink </a>
+   <a href="#" stlye={{textAlign: 'right'}}> GitHubLink </a>
 
   );
 }
