+++ client/src/components/ProfileView/Portfolio/index.jsx
@@ -2,7 +2,7 @@ import React from "react";
 
 const Portfolio = () => {
   return (
-    <a href="" stlye={{textAlign: 'right'}}> Portfolio </a>
+    <a href="#" stlye={{textAlign: 'right'}}> Portfolio </a>
   );
 }
