+++ client/src/components/ProfileView/Resume/index.jsx
@@ -3,7 +3,7 @@ import React from "react";
 
 const Resume = () => {
   return (
-    <a href="" stlye={{textAlign: 'right', marginTop: "10px"}}> Resume </a>
+    <a href="#" stlye={{textAlign: 'right', marginTop: "10px"}}> Resume </a>
   );
 }
