+++ config/config.json
@@ -1,9 +1,9 @@
 {
 	"development": {
-		"username": "o6fm61bpa3oloitc",
-		"password": "ld73vs86fq7q14wb",
-		"database": "g4pmwpgb7l9wck1s",
-		"host": "m7wltxurw8d2n21q.cbetxkdyhwsb.us-east-1.rds.amazonaws.com",
+		"username": "root",
+		"password": "Password12345",
+		"database": "CodeCave",
+		"host": "127.0.0.1",
 		"port": 3306,
 		"dialect": "mysql"
 	},
