+++ models/profile.js
@@ -1,4 +1,3 @@
-const user = require("./user");
 
 module.exports = function (sequelize, DataTypes) {
     var Profile = sequelize.define("Profile", {
