+++ models/user.js.break
@@ -1,0 +1,34 @@
+const Sequelize = require('sequelize')
+const db = require('../database/db.js')
+module.exports = db.sequelize.define(
+    'user',
+    {
+        id: {
+            type: Sequelize.INTEGER,
+            primaryKey: true,
+            autoIncrement: true
+        },
+        first_name: {
+            type: Sequelize.STRING
+        },
+        last_name: {
+            type: Sequelize.STRING
+        },
+        user_name: {
+            type: Sequelize.STRING
+        },
+        email: {
+            type: Sequelize.STRING
+        },
+        password: {
+            type: Sequelize.STRING
+        },
+        created: {
+            type: Sequelize.DATE,
+            defaultValue: Sequelize.NOW
+        }
+    },
+    {
+        timestamps: false
+    }
+)
