+++ package.json
@@ -43,6 +43,7 @@
     "config.json": "0.0.4",
     "cors": "^2.8.5",
     "dotenv": "^8.2.0",
+    "env": "0.0.2",
     "express": "^4.17.1",
     "express-session": "^1.17.1",
     "if-env": "^1.0.4",
